moved genre creation to admin namespace and added test in the feature for a failed genre creation
