+++ app/views/genres/index.html.erb
@@ -1,4 +1,4 @@
-<%= form_for(:genre, url: genres_path, method: :post) do |f| %>
+<%= form_for(:genre, url: admin_genres_path, method: :post) do |f| %>
   <%= f.label :name %>
   <%= f.text_field :name %>
   <%= f.submit 'Create Genre' %>
